+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.12.1 - [Diff](https://github.com/frontfoot/overlay_me/compare/v0.12.0...v0.12.1)
 - renamed embeded jQuery to OMjQuery (and $o accessor)
-- a bit of styling (sizes, fixed on right, onmouseover)
+- removed images from the DOM if not checked (removed 'hide inactives' option)
+- a bit of styling (sizes, fixed on right, onmouseovers)
 
 ### 0.12.0 - [Diff](https://github.com/frontfoot/overlay_me/compare/v0.11.1...v0.12.0)
 - changed the file structure, gem now includes only the minified javascript